将trainin_set按照class分成19份，分别从这19份数据中提取feature
